Add example of an aggregate operation requiring extra data
Implemented a (deliberately silly) business rule that requires
an extra piece of data not available directly from command or state.

This has caused the generic
`handle InventoryState -> InventoryCommand -> InventoryState` to move
from Domain to Application, to be able to call application service
that retrieves the required data.

This also caused the need to pass around more application services
dependencies - before that it was only EventStore port;
for that there is now a structure `ApplicationEnvironment`
that holds such dependencies. Similar approach was already used in
Projections with `DomainEventHandlerContext`.